Add files via upload
A robotic device that will navigate itself in a farm and identify diseased and/or pest affected plants and notify the user about the health of the crop immediately. 
It will also be able to identify rodents and other pests and emit high ultrasonic frequencies to repel them without harming them.
It helps in providing a safe and healthy environment for the optimum growth of crops.